Added :external option for Rcov generator configuration.
MetricFu.rcov[:external] is now can be set through configuration. You can set a pregenerated rcov analysis file as external input and metric_fu will not run your tests / specs / features again.

You can now simply generate and aggregate rcov data from cucumber, rspec or any other source through rake tasks and process the results with MetricFu.

+++ lib/generators/rcov.rb
@@ -19,18 +19,21 @@ module MetricFu
     end
 
     def emit
-      FileUtils.rm_rf(MetricFu::Rcov.metric_directory, :verbose => false)
-      Dir.mkdir(MetricFu::Rcov.metric_directory)
-      test_files = FileList[*MetricFu.rcov[:test_files]].join(' ')
-      rcov_opts = MetricFu.rcov[:rcov_opts].join(' ')
-      output = ">> #{MetricFu::Rcov.metric_directory}/rcov.txt"
-      puts "** Running the specs/tests in the [#{MetricFu.rcov[:environment]}] environment"
-      `RAILS_ENV=#{MetricFu.rcov[:environment]} rcov #{test_files} #{rcov_opts} #{output}`
+      unless MetricFu.rcov[:external]
+        FileUtils.rm_rf(MetricFu::Rcov.metric_directory, :verbose => false)
+        Dir.mkdir(MetricFu::Rcov.metric_directory)
+        test_files = FileList[*MetricFu.rcov[:test_files]].join(' ')
+        rcov_opts = MetricFu.rcov[:rcov_opts].join(' ')
+        output = ">> #{MetricFu::Rcov.metric_directory}/rcov.txt"
+        puts "** Running the specs/tests in the [#{MetricFu.rcov[:environment]}] environment"
+        `RAILS_ENV=#{MetricFu.rcov[:environment]} rcov #{test_files} #{rcov_opts} #{output}`
+      end
     end
 
 
     def analyze
-      output = File.open(MetricFu::Rcov.metric_directory + '/rcov.txt').read
+      output_file = MetricFu.rcov[:external] ? MetricFu.rcov[:external] : MetricFu::Rcov.metric_directory + '/rcov.txt'
+      output = File.open(output_file).read
       output = output.split(NEW_FILE_MARKER)
       
       output.shift # Throw away the first entry - it's the execution time etc.
